Introduce SubmitTransactionFragment and ViewModel for shared code SendFunds/ErgoPay #79
